% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{fetch_questions_from_url}
\alias{fetch_questions_from_url}
\title{Fetch data on written questions as a tibble using a written questions
endpoint URL}
\usage{
fetch_questions_from_url(url, summary = TRUE, take)
}
\arguments{
\item{url}{A valid URL requesting data from the written questions endpoint.}

\item{summary}{A boolean indicating whether to exclude nested and empty
columns in the results. The default is TRUE.}

\item{take}{The number of items to take from the API.}
}
\description{
\code{fetch_questions_from_url} fetches data on written questions from an
endpoint URL and returns it as a tibble containing one row per question.
This internal function allows generic handling of requests for this data
to the written questions endpoint with different URL parameters.
}
\keyword{internal}
