% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fetch_raw_query}
\alias{fetch_raw_query}
\title{Send a query to API and return the result items as a tibble}
\usage{
fetch_raw_query(url)
}
\arguments{
\item{url}{The full API URL specifying the endpoint and request parameters.}
}
\description{
\code{fetch_raw_query} makes an API call to the given endpoint, converts the
results to a tibble, and cleans the column names.
}
\keyword{internal}
