% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\name{fetch_statements_from_url}
\alias{fetch_statements_from_url}
\title{Fetch data on written statements as a tibble using a written statements
endpoint URL}
\usage{
fetch_statements_from_url(url, summary = TRUE, take)
}
\arguments{
\item{url}{A valid URL requesting data from the written statements endpoint.}

\item{summary}{A boolean indicating whether to exclude nested and empty
columns in the results. The default is TRUE.}

\item{take}{The number of items to take from the API.}
}
\description{
\code{fetch_statements_from_url} fetches data on written statements from an
endpoint URL and returns it as a tibble containing one row per statement.
This internal function allows generic handling of requests for this data
to the written statements endpoint with different URL parameters.
}
\keyword{internal}
