% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_wq_variable_names}
\alias{format_wq_variable_names}
\title{Rename written question variable names}
\usage{
format_wq_variable_names(results)
}
\arguments{
\item{results}{The tibble returned from \code{query_results}.}
}
\description{
\code{format_wq_variable_names} takes a tibble returned from \code{query_results}
and renames variables with meaningful descriptions.
}
\keyword{internal}
