% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_ws_variable_types}
\alias{format_ws_variable_types}
\title{Format written statement variable types}
\usage{
format_ws_variable_types(results)
}
\arguments{
\item{results}{The tibble returned from \code{query_results}.}
}
\description{
\code{format_ws_variable_types} takes a tibble returned from \code{query_results}
and specifies variable data types.
}
\keyword{internal}
