% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodies.R
\name{get_answer_bodies}
\alias{get_answer_bodies}
\title{Return the ids and names of Parliamentary question answer bodies}
\usage{
get_answer_bodies()
}
\value{
A tibble with answer body ids and names.
}
\description{
\code{get_answer_bodies} returns the ids and names of bodies responsible
for answer Parliamentary questions. Correct as at 2021-01-26.
}
\details{
id      -- The unique id of an answer body
 body    -- The name of an answer body
}
