% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{missing_argument}
\alias{missing_argument}
\title{Report a missing argument}
\usage{
missing_argument(argument)
}
\arguments{
\item{argument}{The name of the missing argument.}
}
\description{
Report a missing argument
}
\keyword{internal}
