% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\name{process_pagination}
\alias{process_pagination}
\title{Processes hidden pagination of results retrieved from the API}
\usage{
process_pagination(url, take)
}
\arguments{
\item{url}{The full API URL specifying the endpoint and request parameters.}
}
\description{
Processes hidden pagination of results retrieved from the API
}
\keyword{internal}
