% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Send a query to API and return the response as a named list}
\usage{
query(url, take = NULL, warning = FALSE)
}
\arguments{
\item{url}{The full API URL specifying the endpoint and request parameters.}

\item{take}{The number of items to take from the API.}
}
\description{
\code{query} makes an API call to the given endpoint.
}
\keyword{internal}
