% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{query_error}
\alias{query_error}
\title{Report an error with a http response}
\usage{
query_error(code)
}
\arguments{
\item{code}{The status code of the server response}
}
\description{
Report an error with a http response
}
\keyword{internal}
