% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{set_house}
\alias{set_house}
\title{Set house argument based on input}
\usage{
set_house(house = NULL)
}
\arguments{
\item{house}{A string indicating Commons or Lords. Possible values include
"c", "C", "Commons", "l", "L", "Lords". Any other value will default to a
bicameral request.}
}
\description{
\code{set_house} Sets the \code{house} argument based on input.
}
\keyword{internal}
