% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_response_items}
\alias{get_response_items}
\title{Get the data items from an API response as a tibble}
\usage{
get_response_items(response)
}
\arguments{
\item{response}{The response returned from a call to \code{fetch_query}.}
}
\description{
Get the data items from an API response as a tibble
}
\keyword{internal}
